+++ src/index.ts
@@ -1,21 +1,44 @@
 // Copyright (c) Wictor Wilén. All rights reserved.
 // Licensed under the MIT license.
 
-import { BotMessagePreviewActionType, Middleware, TaskModuleMessageResponse, TurnContext } from "botbuilder-core";
 import {
-    ActivityTypes,
+    Middleware, TaskModuleMessageResponse, TurnContext,
     AppBasedLinkQuery,
     MessagingExtensionAction,
     MessagingExtensionQuery,
     MessagingExtensionResult,
     TaskModuleContinueResponse,
+    AdaptiveCardInvokeAction
 } from "botbuilder-core";
+
 import { debug } from "debug";
 
 // Initialize debug logging module
 const log = debug("msteams");
 
-// tslint:disable: max-line-length
+export interface AdaptiveCardRequestValue {
+    action: AdaptiveCardInvokeAction;
+    trigger: "automatic" | "manual";
+}
+
+export interface AdaptiveCardResponseAdaptiveCardBody {
+    statusCode: 200;
+    type: "application/vnd.microsoft.card.adaptive",
+    value: Record<string, unknown>;
+}
+export interface AdaptiveCardResponseMessageBody {
+    statusCode: 200;
+    type: "application/vnd.microsoft.activity.message",
+    value: string;
+}
+
+export interface AdaptiveCardResponseInvalidRequestBody {
+    statusCode: 400;
+    type: "application/vnd.microsoft.error",
+    value: Error;
+}
+
+export declare type AdaptiveCardResponseBody = AdaptiveCardResponseAdaptiveCardBody | AdaptiveCardResponseMessageBody | AdaptiveCardResponseInvalidRequestBody;
 
 /**
  * Defines the processor for the Messaging Extension Middleware
@@ -79,7 +102,7 @@ export interface IMessagingExtensionMiddlewareProcessor {
      * Handles Action.Submit from adaptive cards
      *
      * Note: this is experimental and it does not filter on the commandId which means that if there are
-     * multiple registered message extension processors all will recieve this command. You should ensure to
+     * multiple registered message extension processors all will receive this command. You should ensure to
      * add a specific identifier to your adaptivecard.
      * @param context the turn context
      * @param value the card data
@@ -91,13 +114,39 @@ export interface IMessagingExtensionMiddlewareProcessor {
      * Handles when an item is selected from the result list
      *
      * Note: this is experimental and it does not filter on the commandId which means that if there are
-     * multiple registered message extension processors all will recieve this command. You should ensure to
+     * multiple registered message extension processors all will receive this command. You should ensure to
      * add a specific identifier to your invoke action.
      * @param context the turn context
      * @param value object passed in to invoke action
      * @returns {Promise<MessagingExtensionResult>}
      */
     onSelectItem?(context: TurnContext, value: any): Promise<MessagingExtensionResult>;
+
+    /**
+     * Handles Universal Bot Actions
+     * Examples:
+     * public async onActionExecute(context: TurnContext, value: AdaptiveCardRequestValue): Promise<AdaptiveCardResponseBody> {
+     *   return {
+     *     statusCode: 200,
+     *     type: "application/vnd.microsoft.activity.message",
+     *    value: "A message"
+     *  };
+     * }
+     * or
+     * public async onActionExecute(context: TurnContext, value: AdaptiveCardRequestValue): Promise<AdaptiveCardResponseBody> {
+     *  return {
+     *    statusCode: 200,
+     *    type: "application/vnd.microsoft.card.adaptive",
+     *    value: adaptiveCardObject
+     *  };
+    *  }
+     * See https://docs.microsoft.com/en-us/microsoftteams/platform/task-modules-and-cards/cards/universal-actions-for-adaptive-cards/work-with-universal-actions-for-adaptive-cards#adaptivecardaction-invoke-activity
+     * @param context
+     * @param value
+     * @returns
+     */
+    onActionExecute?(context: TurnContext, value: AdaptiveCardRequestValue):
+        Promise<AdaptiveCardResponseBody>;
 }
 
 const INVOKERESPONSE = "invokeResponse";
@@ -113,10 +162,11 @@ export class MessagingExtensionMiddleware implements Middleware {
      *                  or `undefined` to process all incoming requests
      * @param processor The processor
      */
+    // eslint-disable-next-line no-useless-constructor
     public constructor(
         private commandId: string | undefined,
         private processor: IMessagingExtensionMiddlewareProcessor) {
-
+        // nop
     }
 
     /**
@@ -144,18 +194,18 @@ export class MessagingExtensionMiddleware implements Middleware {
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: {
-                                        composeExtension: result,
+                                        composeExtension: result
                                     },
-                                    status: 200,
-                                },
+                                    status: 200
+                                }
                             });
                         } catch (err) {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: err,
-                                    status: 500,
-                                },
+                                    status: 500
+                                }
                             });
                         }
                         return;
@@ -174,22 +224,22 @@ export class MessagingExtensionMiddleware implements Middleware {
                                             suggestedActions: {
                                                 actions: [{
                                                     type: "openApp",
-                                                    ...result,
-                                                }],
+                                                    ...result
+                                                }]
                                             },
-                                            type: "config",
-                                        },
+                                            type: "config"
+                                        }
                                     },
-                                    status: 200,
-                                },
+                                    status: 200
+                                }
                             });
                         } catch (err) {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: err,
-                                    status: 500,
-                                },
+                                    status: 500
+                                }
                             });
                         }
                         return;
@@ -203,16 +253,16 @@ export class MessagingExtensionMiddleware implements Middleware {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
-                                    status: 200,
-                                },
+                                    status: 200
+                                }
                             });
                         } catch (err) {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: err,
-                                    status: 500,
-                                },
+                                    status: 500
+                                }
                             });
                         }
                         return;
@@ -226,18 +276,18 @@ export class MessagingExtensionMiddleware implements Middleware {
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: {
-                                        composeExtension: result,
+                                        composeExtension: result
                                     },
-                                    status: 200,
-                                },
+                                    status: 200
+                                }
                             });
                         } catch (err) {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: err,
-                                    status: 500,
-                                },
+                                    status: 500
+                                }
                             });
                         }
                         return;
@@ -247,47 +297,55 @@ export class MessagingExtensionMiddleware implements Middleware {
                     }
                     break;
                 case "composeExtension/submitAction":
-                    if ((this.commandId === context.activity.value.commandId || this.commandId === undefined) &&
-                        (this.processor.onSubmitAction || this.processor.onBotMessagePreviewEdit || this.processor.onBotMessagePreviewSend)) {
+                    if (this.commandId === context.activity.value.commandId || this.commandId === undefined) {
                         try {
                             let result;
                             let body;
                             switch (context.activity.value.botMessagePreviewAction) {
                                 case "send":
-                                    result = await this.processor.onBotMessagePreviewSend(context, context.activity.value);
-                                    body = result;
+                                    if (this.processor.onBotMessagePreviewSend) {
+                                        result = await this.processor.onBotMessagePreviewSend(context, context.activity.value);
+                                        body = result;
+                                    }
                                     break;
                                 case "edit":
-                                    result = await this.processor.onBotMessagePreviewEdit(context, context.activity.value);
-                                    body = {
-                                        task: result,
-                                    };
+                                    if (this.processor.onBotMessagePreviewEdit) {
+                                        result = await this.processor.onBotMessagePreviewEdit(context, context.activity.value);
+                                        body = {
+                                            task: result
+                                        };
+                                    }
                                     break;
                                 default:
-                                    result = await this.processor.onSubmitAction(context, context.activity.value);
-                                    body = {
-                                        composeExtension: result,
-                                    };
+                                    if (this.processor.onSubmitAction) {
+                                        result = await this.processor.onSubmitAction(context, context.activity.value);
+                                        body = {
+                                            composeExtension: result
+                                        };
+                                    }
                                     break;
                             }
-
-                            context.sendActivity({
-                                type: INVOKERESPONSE,
-                                value: {
-                                    body,
-                                    status: 200,
-                                },
-                            });
+                            if (result) {
+                                context.sendActivity({
+                                    type: INVOKERESPONSE,
+                                    value: {
+                                        body,
+                                        status: 200
+                                    }
+                                });
+                                return;
+                            }
                         } catch (err) {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: err,
-                                    status: 500,
-                                },
+                                    status: 500
+                                }
                             });
+                            return;
                         }
-                        return;
+
                     }
                     break;
                 case "composeExtension/fetchTask":
@@ -297,23 +355,23 @@ export class MessagingExtensionMiddleware implements Middleware {
                         try {
                             const result = await this.processor.onFetchTask(context, context.activity.value);
 
-                            const body = result.type === "continue" || result.type === "message" ?
-                                { task: result } :
-                                { composeExtension: result };
+                            const body = result.type === "continue" || result.type === "message"
+                                ? { task: result }
+                                : { composeExtension: result };
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body,
-                                    status: 200,
-                                },
+                                    status: 200
+                                }
                             });
                         } catch (err) {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: err,
-                                    status: 500,
-                                },
+                                    status: 500
+                                }
                             });
                         }
                         return;
@@ -326,16 +384,16 @@ export class MessagingExtensionMiddleware implements Middleware {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
-                                    status: 200,
-                                },
+                                    status: 200
+                                }
                             });
                         } catch (err) {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: err,
-                                    status: 500,
-                                },
+                                    status: 500
+                                }
                             });
                         }
                     }
@@ -348,24 +406,49 @@ export class MessagingExtensionMiddleware implements Middleware {
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: {
-                                        composeExtension: result,
+                                        composeExtension: result
                                     },
-                                    status: 200,
-                                },
+                                    status: 200
+                                }
                             });
-                            return;
-                            // we're doing a return here and not next() so we're not colliding with
-                            // any botbuilder-teams invoke things. This however will also invalidate the use
-                            // of multiple message extensions using selectItem - only the first one will be triggered
                         } catch (err) {
                             context.sendActivity({
                                 type: INVOKERESPONSE,
                                 value: {
                                     body: err,
-                                    status: 200,
-                                },
+                                    status: 500
+                                }
                             });
                         }
+                        return;
+                        // we're doing a return here and not next() so we're not colliding with
+                        // any botbuilder-teams invoke things. This however will also invalidate the use
+                        // of multiple message extensions using selectItem - only the first one will be triggered
+
+                    }
+                    break;
+                case "adaptiveCard/action":
+                    if (this.processor.onActionExecute) {
+                        try {
+                            const result = await this.processor.onActionExecute(context, context.activity.value);
+
+                            context.sendActivity({
+                                type: INVOKERESPONSE,
+                                value: {
+                                    body: result,
+                                    status: 200
+                                }
+                            });
+                        } catch (err) {
+                            context.sendActivity({
+                                type: INVOKERESPONSE,
+                                value: {
+                                    body: err,
+                                    status: 500
+                                }
+                            });
+                        }
+                        return;
                     }
                     break;
                 default: